requirements.txt
aniso8601==7.0.0
boto3==1.9.224
botocore==1.12.224
Django==2.2.5
django-cors-headers==3.1.0
django-filter==2.2.0
django-graphql-jwt==0.2.2
django-storages==1.7.1
djangorestframework==3.10.3
docutils==0.15.2
graphene==2.1.8
graphene-django==2.5.0
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.9.0
jmespath==0.9.4
Markdown==3.1.1
Pillow==6.1.0
promise==2.2.1
psycopg2==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
Rx==3.0.1
s3transfer==0.2.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
